Move data loading to interface
