+++ BlasModInstaller/Loading/ILoader.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
+﻿
 namespace BlasModInstaller.Loading
 {
     internal interface ILoader
